design changes in feed & detail post, added buttons to post & detailpost, fixed: no background pictures anymore
