+++ CoolGag/client/detailedPost.css
@@ -1,0 +1,24 @@
+
+.author-tag{
+	float: right;
+}
+
+.detailPost-view{
+  overflow: auto;
+  border-radius: 10px;
+  background: rgba(204, 204, 204, 0.03);
+  padding: 10px;
+  display: block;
+  margin-top: 10px;
+
+}
+
+.post-img {
+	display: block;
+ 	margin: 10px auto 10px auto;
+}
+
+.singlepost-container{
+	display: block;
+	margin: 0 auto;
+}
