+++ CoolGag/imports/ui/components/ShowPost.js
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 
     return (
       <div className='w-100 flex justify-center'>
-        <div className='w-100' style={{ maxWidth: 400 }}>
+        <div className='w-100' style={{ maxWidth: 600 }}>
         {this.props.data.allPosts.map((post) =>
           {
             if (post.id == this.props.params.id){
