+++ CoolGag/package.json
@@ -15,7 +15,7 @@
     "graphql": "^0.10.1",
     "graphql-faker": "^1.2.1",
     "graphql-tag": "^1.3.2",
-    "html2canvas": "0.5.0-beta4",
+    "html2canvas": "^0.5.0-beta4",
     "meteor-node-stubs": "^0.2.11",
     "minimatch": "^3.0.4",
     "prop-types": "^15.5.10",
@@ -32,7 +32,7 @@
     "react-uikit-comment": "^2.0.2",
     "react-uikit-comment-list": "^2.0.1",
     "reactstrap": "^4.6.2",
-    "request": "^2.67.0",
+    "request": "^2.81.0",
     "tachyons": "^4.7.4",
     "url": "0.11.0"
   },
